[SEC-CARD01] - Implementação de CORS e proteção de rotas por token e roles
