+++ src/main/resources/application-dev.yml~
@@ -1,8 +1,4 @@
 quarkus:
-  hibernate-orm:
-    package: org.ged
-    database:
-      generation: drop-and-create
   server:
     port: 8080
     servlet:
@@ -28,17 +24,25 @@ quarkus:
     console:
       level: ALL
   http:
-    root-path: /ged/api
     cors:
-      ~: true
-      origins: http://localhost:4200
       headers:
         Access-Control-Allow-Origin: true
-        Access-Control-Allow-Methods: GET, POST, PUT, DELETE, OPTIONS
-        Access-Control-Allow-Headers: accept, content-type, x-requested-with, authorization, x-custom
-        expose-headers: Content-Disposition
+        Access-Control-Allow-Methods: GET, POST, PUT, DELETE
+        Access-Control-Allow-Headers: accept, content-type, x-requested-with, authorization
         Access-Control-Allow-Content-Type: application/json
         Access-Control-Allow-Credentials: true
+
+      ~: true
+      origins:
+       - http://localhost:4200
+       - http://localhost:8080
+       - http://localhost:9081
+    root-path: /ged/api
+
+  hibernate-orm:
+    package: org.ged
+    database:
+      generation: drop-and-create
 
 #  flyway:
 #    migrate-at-start: true
